+++ backend/app/middleware/handlerErrors.js
@@ -1,9 +1,9 @@
 
-const logger = require("../../pino-loger/logger")
+const logger = require("../../logger")
 
 //No route found handler
 const routeFoundHandler = (req, res, next) => {
-	logger.fatal("Route not found")
+	logger.error("Route not found")
 	res.status(404);
 	res.json({
 		message: "Error. Route not found",
@@ -15,7 +15,7 @@ const routeFoundHandler = (req, res, next) => {
 const errorHandler = (err, req, res, next) => {
 
 	const {statusCode = 500, code, header, message} = err;
-	logger.fatal({code, header, message})
+	logger.error({code, header, message})
 	res.status(statusCode);
 	res.json({
 		code,
