+++ backend/package.json
@@ -20,7 +20,6 @@
     "argon2": "0.28.2",
     "cors": "2.8.5",
     "csvtojson": "^2.0.10",
-    "detect-browser": "^5.3.0",
     "dotenv": "10.0.0",
     "express": "4.17.1",
     "express-jsdoc-swagger": "1.6.2",
